+++ CHANGELOG.md
@@ -1,13 +1,57 @@
 # Changelog
 
-### [0.1.2](https://github.com/googleapis/python-video-live-stream/compare/v0.1.1...v0.1.2) (2022-04-28)
+## [0.2.0](https://github.com/googleapis/python-video-live-stream/compare/v0.1.5...v0.2.0) (2022-07-16)
+
+
+### Features
+
+* add audience parameter ([f8824ce](https://github.com/googleapis/python-video-live-stream/commit/f8824ce3ca3055939a6d2ce407c414e21e140e4b))
+
+
+### Bug Fixes
+
+* **deps:** require google-api-core>=1.32.0,>=2.8.0 ([#78](https://github.com/googleapis/python-video-live-stream/issues/78)) ([e2f1b17](https://github.com/googleapis/python-video-live-stream/commit/e2f1b17565d44f34b33d8881243fb683fba3f37f))
+* require python 3.7+ ([#77](https://github.com/googleapis/python-video-live-stream/issues/77)) ([a4e7d42](https://github.com/googleapis/python-video-live-stream/commit/a4e7d427c504ac1f39e64f7417731832926a2500))
+
+
+### Documentation
+
+* align channel config with best practices ([#71](https://github.com/googleapis/python-video-live-stream/issues/71)) ([2a1a5b2](https://github.com/googleapis/python-video-live-stream/commit/2a1a5b20f6ecb9530fa48ba22f384d266301a825))
+
+## [0.1.5](https://github.com/googleapis/python-video-live-stream/compare/v0.1.4...v0.1.5) (2022-06-03)
+
+
+### Bug Fixes
+
+* **deps:** require protobuf <4.0.0dev ([#65](https://github.com/googleapis/python-video-live-stream/issues/65)) ([a072bbf](https://github.com/googleapis/python-video-live-stream/commit/a072bbf7d411b4595e5f915866a839eb52e3d208))
+
+
+### Documentation
+
+* fix changelog header to consistent size ([#66](https://github.com/googleapis/python-video-live-stream/issues/66)) ([fbc3528](https://github.com/googleapis/python-video-live-stream/commit/fbc3528b4d2f22751c0f5a99bae513668fcb3db5))
+
+## [0.1.4](https://github.com/googleapis/python-video-live-stream/compare/v0.1.3...v0.1.4) (2022-05-22)
+
+
+### Documentation
+
+* updated comments to match API behaviors ([#45](https://github.com/googleapis/python-video-live-stream/issues/45)) ([bdc9463](https://github.com/googleapis/python-video-live-stream/commit/bdc9463156a63439ef41621405394042d01f3cb9))
+
+## [0.1.3](https://github.com/googleapis/python-video-live-stream/compare/v0.1.2...v0.1.3) (2022-05-03)
+
+
+### Documentation
+
+* **samples:** add sample for creating a channel with a failover backup input ([#37](https://github.com/googleapis/python-video-live-stream/issues/37)) ([9587382](https://github.com/googleapis/python-video-live-stream/commit/958738237c7c055a3a656ce2c3a7964e49d6b6b1))
+
+## [0.1.2](https://github.com/googleapis/python-video-live-stream/compare/v0.1.1...v0.1.2) (2022-04-28)
 
 
 ### Documentation
 
 * **samples:** add code samples ([#21](https://github.com/googleapis/python-video-live-stream/issues/21)) ([484271d](https://github.com/googleapis/python-video-live-stream/commit/484271de8790e2d9d0d4ea02feb344398317ffe5))
 
-### [0.1.1](https://github.com/googleapis/python-video-live-stream/compare/v0.1.0...v0.1.1) (2022-03-05)
+## [0.1.1](https://github.com/googleapis/python-video-live-stream/compare/v0.1.0...v0.1.1) (2022-03-05)
 
 
 ### Bug Fixes